requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
pandas==2.0.3
geopandas==0.13.2
plotly==5.17.0
folium==0.14.0
python-dotenv==1.0.0
SQLAlchemy==2.0.21
marshmallow==3.20.1
